kogniterm/requirements.txt
python-dotenv
langgraph
prompt_toolkit
langchain_mcp_adapters
langchain_community
bs4
PyGithub
rich
matplotlib
google-generativeai
jupyter_client
loguru
colorama
gitignore-parser
watchdog
litellm

GitPython